# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi REST API Rumah Sakit",
        "description": "Dokumentasi API v1.0.0",
        "contact": {
            "email": "prasdimasetyo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/patient": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Mengambil daftar semua pasien",
                "description": "Endpoint ini memerlukan autentikasi dengan menggunakan Bearer Token. Harap sertakan token bearer dalam header 'Authorization' saat melakukan permintaan.",
                "operationId": "051dba4eed49b0cf2377a73e66aa50d4",
                "responses": {
                    "200": {
                        "description": "Daftar pasien berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "patients": [
                                        {
                                            "id": 1,
                                            "name": "John Doe"
                                        },
                                        {
                                            "id": 2,
                                            "name": "Jane Smith"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Membuat data pasien baru",
                "description": "Endpoint ini memerlukan autentikasi dengan menggunakan Bearer Token. Harap sertakan token bearer dalam header 'Authorization' saat melakukan permintaan.",
                "operationId": "bd276a8576aab17838c1ae2d423f333d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "John Doe",
                                "email": "john.doe@example.com",
                                "phone": "1234567890",
                                "age": 30,
                                "gender": "Male",
                                "sickness": "Flu"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pasien berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "msg": "Berhasil Dibuat"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "fail",
                                    "message": [
                                        "The name field is required.",
                                        "The email field is required."
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal membuat data pasien",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak dapat membuat data pasien"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patient/{id}": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Menampilkan detail pasien berdasarkan ID",
                "description": "Endpoint ini memerlukan autentikasi dengan menggunakan Bearer Token. Harap sertakan token bearer dalam header 'Authorization' saat melakukan permintaan.",
                "operationId": "3e4f6fe1198e4341e2095d34750020b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari pasien yang ingin ditampilkan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pasien berhasil ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "patient": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "phone": "1234567890",
                                        "age": 30,
                                        "gender": "male",
                                        "sickness": "Flu"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Pasien tidak ditemukan"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal menampilkan data pasien",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak dapat menampilkan data pasien"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Patients"
                ],
                "summary": "Memperbarui data pasien berdasarkan ID",
                "description": "Endpoint ini memerlukan autentikasi dengan menggunakan Bearer Token. Harap sertakan token bearer dalam header 'Authorization' saat melakukan permintaan.",
                "operationId": "9c6987fcff28b4c6d19e755da59e7293",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari pasien yang ingin diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "John Doe",
                                "email": "john@example.com",
                                "phone": "1234567890",
                                "age": 30,
                                "gender": "male",
                                "sickness": "Flu"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pasien berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "Pasien berhasil diperbarui"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Pasien tidak ditemukan"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal memperbarui data pasien",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak dapat memperbarui data pasien"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Patients"
                ],
                "summary": "Menghapus data pasien berdasarkan ID",
                "description": "Endpoint ini memerlukan autentikasi dengan menggunakan Bearer Token. Harap sertakan token bearer dalam header 'Authorization' saat melakukan permintaan.",
                "operationId": "6d0d4ab9f1b092141ed19c02b557553f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari pasien yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pasien berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "Pasien berhasil dihapus"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Pasien tidak ditemukan"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gagal menghapus data pasien",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak dapat menghapus data pasien"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}